pyproject.toml
[project]
name = "doc-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "ipykernel>=6.29.5",
    "openai-agents>=0.1.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
]
